update when rerun
not faster but if you run it again, on same repositories folder, it will give you at the end a list of all updated. so not extract files with same version
he still have to search all repos so it's quite long. but well it's not so long running behind. but maybe my modifications can be a base to make it faster